Develop

+++ download-manager/src/main/java/me/henrytao/downloadmanager/internal/FileUtils.java
@@ -33,8 +33,22 @@ import java.security.MessageDigest;
 
 class FileUtils {
 
-  public static boolean delete(File file) {
-    return file == null || (file.exists() && file.delete());
+  public static void delete(File fileOrDirectory) {
+    if (fileOrDirectory == null || !fileOrDirectory.exists()) {
+      return;
+    }
+    if (fileOrDirectory.isDirectory()) {
+      for (File file : fileOrDirectory.listFiles()) {
+        delete(file);
+      }
+      if (!fileOrDirectory.delete()) {
+        fileOrDirectory.deleteOnExit();
+      }
+    } else {
+      if (!fileOrDirectory.delete()) {
+        fileOrDirectory.deleteOnExit();
+      }
+    }
   }
 
   @SuppressWarnings("ResultOfMethodCallIgnored")
@@ -81,26 +95,28 @@ class FileUtils {
     if (input == null || !input.exists() || output == null) {
       throw new IllegalArgumentException("Input and Output files can not be null");
     }
-    output.getParentFile().mkdirs();
-    if (output.exists() && autoRename) {
-      output = autoRenameIfExists(output);
-    }
-    FileChannel inputChannel = null;
-    FileChannel outputChannel = null;
-    try {
-      inputChannel = new FileInputStream(input).getChannel();
-      outputChannel = new FileOutputStream(output).getChannel();
-      inputChannel.transferTo(0, inputChannel.size(), outputChannel);
-      inputChannel.close();
-      input.delete();
-    } finally {
-      if (inputChannel != null) {
-        //noinspection ThrowFromFinallyBlock
-        inputChannel.close();
+    if (!input.equals(output)) {
+      output.getParentFile().mkdirs();
+      if (output.exists() && autoRename) {
+        output = autoRenameIfExists(output);
       }
-      if (outputChannel != null) {
-        //noinspection ThrowFromFinallyBlock
-        outputChannel.close();
+      FileChannel inputChannel = null;
+      FileChannel outputChannel = null;
+      try {
+        inputChannel = new FileInputStream(input).getChannel();
+        outputChannel = new FileOutputStream(output).getChannel();
+        inputChannel.transferTo(0, inputChannel.size(), outputChannel);
+        inputChannel.close();
+        input.delete();
+      } finally {
+        if (inputChannel != null) {
+          //noinspection ThrowFromFinallyBlock
+          inputChannel.close();
+        }
+        if (outputChannel != null) {
+          //noinspection ThrowFromFinallyBlock
+          outputChannel.close();
+        }
       }
     }
     return output;
